refactor: create config files
